Replacing references to reader_impl_helpers with column_path_helpers in multiple files. Local complie pass, now trigger CI
